refactor source to be in different files according to functionality
