+++ src/errors.def
@@ -1,0 +1,6 @@
+No error.
+Bad CLI input.
+Could not open file for reading.
+Function could not execute properly.
+Could not retrieve axes data.
+Could not read data from device.
